fix gutter partial modification indicator
When only deleted lines of a modification are staged, gutter should
show deletion and add markers so that deletion marker is colored
as in index and addition marker is shown as not in index.

This change takes diffs between index and head and uses that
to identify this situation. Uses class git-index-partial to mark the line
where deletion marker should be placed.