Enhance index change detection
Detecting status changes does not work when we need to
trigger updates whenever index changes. When first hunk
is added to index its status changes, but adding more hunks
does not affect status. And the other way round.

fs.watch() is used to follow index, but it is not sufficient either.
Index file does not change when last commit is reverted to index
(git reset --soft HEAD~)

repository.onDidChangeStatuses() solves this but it does not
say which paths are modified. For that reason we keep the
value in @status variable to detect changes. This way we can
prevent updating unnecessarily all open editors.

Also we now use async version to get the repository.
Sync version is deprecated. Diffs update only after repository is
fetched, and all git related operations and movements fail silently
during the fetch.